restored byte[0] test and moved IMsgHeader to modP2p
