+++ modP2p/src/org/aion/p2p/Header.java
@@ -27,7 +27,6 @@ package org.aion.p2p;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import org.aion.base.type.IMsgHeader;
 
 /** @author chris */
 public final class Header implements IMsgHeader {
